Make it easier to find the direct preceding version
This is a trivial change that stores the UID of the object this object was
versioned from in the same annotation as the rest of the information. This
means you do not have to hit the catalog to find that information, although
you still need to hit the reference catalog to get the object.